refactor: FeedFinder -> FeedSelector 명칭 변경 및 일부 타입 변환

+++ backend/src/main/java/com/example/backend/social/feed/dto/FeedResponse.java
@@ -23,8 +23,8 @@ public class FeedResponse {
 	private Long postId;
 	private List<String> imgUrlList;
 	private String content;
-	private int likesCount;
-	private int commentCount;
+	private Long likesCount;
+	private Long commentCount;
 	private LocalDateTime createdDate;
 
 	// 헤시태그 정보
